backend/requirements.txt
# ============================================
# FILE 1: backend/requirements.txt
# ============================================

# Web Framework
fastapi==0.103.0
uvicorn[standard]==0.23.2
python-multipart==0.0.6

# Database
sqlalchemy==2.0.20
psycopg2-binary==2.9.11
alembic==1.12.0

# Redis Cache
redis==5.0.0
hiredis==2.2.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==41.0.4

# AI/ML (for AI engine integration)
openai==0.28.0
transformers==4.33.0
torch==2.9.0
numpy==1.24.3

# HTTP Clients
httpx==0.24.1
aiohttp==3.8.5

# Validation
pydantic==2.3.0
pydantic-settings==2.0.3
email-validator==2.0.0

# Monitoring & Logging
prometheus-client==0.17.1
python-json-logger==2.0.7

# Rate Limiting
slowapi==0.1.9

# CORS
# python-cors==1.0.0


# Testing
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.24.1

# Development
black==23.9.1
flake8==6.1.0
mypy==1.5.1